+++ R/checkParents.R
@@ -26,7 +26,7 @@ checkParentIDs <- function(ped, verbose = FALSE, repair = FALSE,
                            parentswithoutrow = repair) {
   # Standardize column names in the input dataframe
   ped_og <- ped
-  ped <- standardizeColnames(ped)
+  ped <- standardizeColnames(ped, verbose = verbose)
 
 
   # Initialize a list to store validation results
@@ -299,4 +299,49 @@ checkParentIDs <- function(ped, verbose = FALSE, repair = FALSE,
 #' @return A corrected pedigree
 repairParentIDs <- function(ped, verbose = FALSE) {
   checkParentIDs(ped = ped, verbose = verbose, repair = TRUE)
+}
+
+#' Add Phantom Parents
+#'
+#' This function adds phantom parents to a pedigree.
+#' @inheritParams checkParentIDs
+
+addPhantoms <- function(ped, verbose, pid, validation_results) {
+  # Add parents who appear in momID or dadID but are missing from ID
+  new_entries <- data.frame()
+
+  listed_parents <- unique(c(ped$momID, ped$dadID))
+  listed_parents <- listed_parents[!is.na(listed_parents)]
+
+  existing_ids <- ped$ID
+  missing_parents <- setdiff(listed_parents, existing_ids)
+
+  if (length(missing_parents) > 0) {
+    if (verbose) {
+      cat("Adding parents who were listed in momID/dadID but missing from ID:\n")
+      print(missing_parents)
+    }
+
+    for (pid in missing_parents) {
+      role <- unique(
+        c(
+          if (pid %in% ped$momID) "mom" else NULL,
+          if (pid %in% ped$dadID) "dad" else NULL
+        )
+      )
+      inferred_sex <- if ("mom" %in% role) validation_results$female_var else validation_results$male_var
+
+      new_row <- ped[1, ]
+      new_row$ID <- pid
+      new_row$dadID <- NA
+      new_row$momID <- NA
+      new_row$sex <- inferred_sex
+      new_entries <- rbind(new_entries, new_row)
+    }
+  }
+  ped <- merge(ped, new_entries, all = TRUE)
+  if (verbose) {
+    cat("Added phantom parents for:", paste(new_entries$ID, collapse = ", "), "\n")
+  }
+  return(ped)
 }